+++ sheets/tasks.py
@@ -4,6 +4,8 @@ from googleapiclient.errors import HttpError
 from mitxpro.celery import app
 from mitxpro.utils import now_in_utc
 from sheets import api
+import sheets.coupon_assign_api
+import sheets.coupon_request_api
 from sheets.constants import ASSIGNMENT_SHEET_ENROLLED_STATUS
 
 
@@ -14,7 +16,7 @@ def handle_unprocessed_coupon_requests():
     coupons, updates the request sheet to indicate that it was processed, and creates
     the necessary coupon assignment sheets.
     """
-    coupon_request_handler = api.CouponRequestHandler()
+    coupon_request_handler = sheets.coupon_request_api.CouponRequestHandler()
     results = coupon_request_handler.process_sheet()
     return results
 
@@ -24,7 +26,7 @@ def handle_incomplete_coupon_assignments():
     """
     Processes all as-yet-incomplete coupon assignment spreadsheets
     """
-    coupon_assignment_handler = api.CouponAssignmentHandler()
+    coupon_assignment_handler = sheets.coupon_assign_api.CouponAssignmentHandler()
     processed_spreadsheet_metadata = (
         coupon_assignment_handler.process_assignment_spreadsheets()
     )
@@ -37,7 +39,7 @@ def update_incomplete_assignment_delivery_statuses():
     Fetches all BulkCouponAssignments that have assignments but have not yet finished delivery, then updates the
     delivery status for each depending on what has been sent.
     """
-    coupon_assignment_handler = api.CouponAssignmentHandler()
+    coupon_assignment_handler = sheets.coupon_assign_api.CouponAssignmentHandler()
     updated_assignments = (
         coupon_assignment_handler.update_incomplete_assignment_message_statuses()
     )
@@ -63,7 +65,7 @@ def set_assignment_rows_to_enrolled(sheet_update_map):
             number of updated assignments in that sheet.
     """
     now = now_in_utc()
-    coupon_assignment_handler = api.CouponAssignmentHandler()
+    coupon_assignment_handler = sheets.coupon_assign_api.CouponAssignmentHandler()
     result_summary = {}
     for sheet_id, assignment_code_email_pairs in sheet_update_map.items():
         # Convert the list of lists into a set of tuples, and set the strings to lowercase so
